Upgrade to Rust edition 2024, fix clippy warnings, and improve code quality
- Upgrade to Rust 2024 edition with proper unsafe blocks in unsafe fns
- Update allocator-api2 0.2 -> 0.4, criterion 0.5 -> 0.8
- Fix all 8 clippy warnings (missing Safety docs, suspicious_open_options,
  is_multiple_of, io_other_error)
- Fix Error::OverAlloc bug in prealloc.rs (variant didn't exist, changed to OverGrow)
- Fix prealloc.rs grow() signature to match RawMem trait
- Move tests from src/ (raw_place.rs, async_mem.rs) to tests/ folder
- Update rustfmt.toml to stable-only options
- Run cargo fmt for consistent edition 2024 formatting
