+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "platform-mem"
 version = "0.1.0-pre+beta.2"
-edition = "2021"
+edition = "2024"
 authors = ["uselesssgoddess", "Linksplatform Team <linksplatformtechnologies@gmail.com>"]
 license = "Unlicense"
 repository = "https://github.com/linksplatform/mem-rs"
@@ -16,7 +16,7 @@ default = []
 async = ["tokio"]
 
 [dependencies]
-allocator-api2 = "0.2"
+allocator-api2 = "0.4"
 memmap2 = "0.9"
 tempfile = "3"
 thiserror = "2"
@@ -28,7 +28,7 @@ tokio = { version = "1", features = ["sync"], optional = true }
 paste = "1"
 quickcheck = "1"
 quickcheck_macros = "1"
-criterion = { version = "0.5", features = ["html_reports"] }
+criterion = { version = "0.8", features = ["html_reports"] }
 tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
 
 [[example]]
