+++ rustfmt.toml
@@ -1,6 +1,2 @@
-error_on_line_overflow = true
-error_on_unformatted = true
-version = "Two"
-
-imports_granularity = "One"
+edition = "2024"
 use_small_heuristics = "Max"
