+++ src/prealloc.rs
@@ -5,6 +5,11 @@ use {
         ops::{Deref, DerefMut},
     },
 };
+
+/// Pre-allocated memory backed by an existing mutable slice or `Vec`.
+///
+/// This implementation does not perform dynamic allocation — it uses
+/// a fixed buffer provided at construction time.
 #[derive(Debug)]
 pub struct PreAlloc<P> {
     place: P,
@@ -12,7 +17,7 @@ pub struct PreAlloc<P> {
 }
 
 impl<T, P: Deref<Target = [T]> + DerefMut> PreAlloc<P> {
-    /// Constructs new `PreAlloc`
+    /// Constructs a new `PreAlloc` wrapping the given buffer.
     pub fn new(place: P) -> Self {
         Self { place, used: 0 }
     }
@@ -32,17 +37,18 @@ impl<T, P: Deref<Target = [T]> + DerefMut> RawMem for PreAlloc<P> {
     unsafe fn grow(
         &mut self,
         addition: usize,
-        fill: impl FnOnce(&mut [MaybeUninit<Self::Item>]),
+        fill: impl FnOnce(usize, (&mut [Self::Item], &mut [MaybeUninit<Self::Item>])),
     ) -> Result<&mut [Self::Item]> {
         let cap = self.used.checked_add(addition).ok_or(CapacityOverflow)?;
         let available = self.place.len();
 
         if let Some(slice) = self.place.get_mut(self.used..cap) {
-            fill(mem::transmute(&mut slice[..]));
+            let uninit = unsafe { mem::transmute::<&mut [T], &mut [MaybeUninit<T>]>(&mut slice[..]) };
+            fill(0, (&mut [], uninit));
             self.used = cap;
             Ok(slice)
         } else {
-            Err(crate::Error::OverAlloc { available, to_alloc: cap })
+            Err(crate::Error::OverGrow { available, to_grow: cap })
         }
     }
